DB1/requirements.txt
# amqp==5.2.0
# archspec==0.2.3
# asgiref==3.8.1
# attrs==24.2.0
# babel==2.11.0
# billiard==4.2.1
# boltons==23.0.0
# # brotli-python==1.0.9
# # bzip2==1.0.8
# # ca-certificates==2024.9.24
# celery==5.4.0
# certifi==2024.8.30
# cffi==1.16.0
# charset-normalizer==2.0.4
# click==8.1.7
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# colorama==0.4.6
# # conda==24.9.2
# # conda-content-trust==0.2.0
# # conda-libmamba-solver==24.1.0
# # conda-package-handling==2.3.0
# # conda-package-streaming==0.10.0
# # console_shortcut_miniconda==0.1.1
# cryptography==42.0.5
# distro==1.9.0
django==5.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
# expat==2.6.2
# flower==0.9.7
# fmt==9.1.0
frozendict==2.4.2
greenlet==3.0.1
humanize==3.10.0
idna==3.7
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
pika==1.3.2
pip==24.0
platformdirs==3.10.0
pluggy==1.0.0
referencing==0.35.1
# reproc==14.2.4
# reproc-cpp==14.2.4
requests==2.32.3
rpds-py==0.20.0
ruamel.yaml==0.17.21
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlalchemy==2.0.34
# sqlite==3.45.3
sqlparse==0.5.1
# tk==8.6.14
tornado==6.4.1
tqdm==4.66.4
truststore==0.8.0
typing-extensions==4.11.0
typing_extensions==4.11.0
tzdata==2024.1
# ucrt==10.0.22621.0
uritemplate==4.1.1
urllib3==2.2.2
# vc==14.2
# vc14_runtime==14.40.33810
# vine==5.1.0
# vs2015_runtime==14.40.33810
wcwidth==0.2.13
wheel==0.43.0
win_inet_pton==1.1.0
# xz==5.4.6
# yaml-cpp==0.8.0
# zlib==1.2.13
zstandard==0.22.0
# zstd==1.5.5
